backend/requirements.txt
bleach==5.0.1
Django==4.1.2
django-cors-headers==3.13.0
django-filter==22.1
django-location-field==2.1.0
django-rest-knox==4.2.0
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
googlemaps==4.7.0
Pillow==9.3.0
pycparser==2.21
Pygments==2.13.0
python-dateutil==2.8.2
requests==2.28.1
